+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - console: bug fixes for RS-to-RS relationships
 - console: allow users to remove prefix / suffix / root field namespace from a remote schema
 - console: new "add remote schema" page (with GQL customization)
+- console: fix console crash on adding pg sources with connection params through api
 - cli: avoid exporting hasura-specific schemas during hasura init (#8352)
 - cli: fix performance regression in `migrate status` command (fix #8398)